Move IsStaffMixin to info_pages.
